% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_arid.R
\docType{data}
\name{arid}
\alias{arid}
\title{A list of binary matrices and their associated parameters}
\format{A list of matrices, representing values of a black and white 
  aerial picture.}
\source{
Aerial images of a stressed ecosystem, with an increasing aridity
  from first to last matrix
}
\usage{
arid
}
\description{
A list of binary matrices and their associated parameters
}
\references{
TODO
}
\keyword{datasets}

