% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_binary_matrix.R
\name{as.binary_matrix}
\alias{as.binary_matrix}
\title{Coerce to a binary matrix}
\usage{
as.binary_matrix(x, state = NA)
}
\arguments{
\item{x}{Any R object}

\item{state}{A reference state to use when converting to a binary matrix}
}
\description{
Coerce to a binary matrix
}
\details{
`as.binary_matrix` is a generic function that tries to convert
  object to binary matrices. It can be applied to lists of objects: in this
  case it will try to convert each object individually and return a list.
}
\examples{
# Matrix example: convert all cells to binary data using state 3 as reference
mat <- matrix(sample(c(1,2,3), 100, replace = TRUE), nrow = 10, ncol = 10)
as.binary_matrix(mat, state = 3)
}

