% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarse_graining.R
\name{coarse_grain}
\alias{coarse_grain}
\title{Matrix coarse-graining}
\usage{
coarse_grain(mat, subsize)
}
\arguments{
\item{mat}{A square matrix}

\item{subsize}{Dimention of the submatrix. This has to be a positive integer smaller than the dimention of input matrix.}
}
\value{
A square matrix of reduced dimention.
}
\description{
This function averages the spatial data locally. 
It divides the input matrix into submatrices of dimention subsize 
and averages the spatial data in these submatrices.
By doing this, the dimention of resultant matrix is reduced by a factor of subsize.
}
\details{
If the data is classified into discrete units, 
the calculation of variance and skewness can give spurious results
irrelevant to the proximity to transition. Therefore, discrete data should be 
'coarse-grained' before calculating the spatial early warning signals. 
However, this can also be applied to continuous state data.
}
\references{
Sankaran, S., Majumder, S., Kefi, S. and Guttal, V. (2017). 
Implications of being discrete and spatial for detecting early warning signals of regime shifts. Ecological Indicators.
}
