% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Indicator_CorrelationFunction.R
\name{CorrF}
\alias{CorrF}
\title{Spatial Warning Indicators: Correlation Function}
\usage{
CorrF(data)
}
\arguments{
\item{data}{is a list of binary matrices of length >=1 (OR) a single binary matrix}
}
\value{
A list of lists (of correlation at different distances, from 0 to SystemSize/2 with increments of 1 unit) for every matrix of input list (OR) a single vector (of correlation at different distances from 0 to SystemSize/2 with increments of 1 unit) if input is a single matrix
}
\description{
MAIN function to calculate correlation function
}
\details{
dependencies include corrfunc.R (included in the repository)
}

