% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corrfunc.R
\name{corrfunc}
\alias{corrfunc}
\title{Function to calculate correlation at different distances.}
\usage{
corrfunc(mat)
}
\arguments{
\item{binary}{square matrix}
}
\value{
Vector of correlation at distances from 0 to System Size/2 with increments of 1 unit.
}
\description{
Operates on a single square matrix
}
\details{
This function is a dependency of Indicator_CorrelationFunction.R and is called by it
}

