% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_spectral_indic_methods.R
\name{extract_spectrum}
\alias{extract_spectrum}
\title{Extract the r-spectrum from objects}
\usage{
extract_spectrum(x, ...)
}
\arguments{
\item{x}{An object produced by \code{spectral_sews} or the result of the 
\code{indictest} function called on such object}

\item{...}{Other arguments are ignored}
}
\value{
The empirical r-spectrum as a \code{data.frame}
}
\description{
Extract the r-spectrum from objects produced by
  \code{spectral_sews}.
}
\examples{

# Extract the r-spectrum after computing indicators
indics <- spectral_sews(serengeti[2:3])
extract_spectrum(indics) 

}
\seealso{
\code{\link{spectral_sews}}, \code{\link{rspectrum}}
}
