% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_flowlength.R
\name{flowlength_sews}
\alias{flowlength_sews}
\title{Flow length indicator}
\usage{
flowlength_sews(mat, slope = 20, cell_size = 1, method = "planar")
}
\arguments{
\item{mat}{The input matrix (must be a logical matrix)}

\item{slope}{The slope of the area documented by the matrix (in degree).}

\item{cell_size}{The horizontal size of a cell in the matrix (as viewed 
from above).}

\item{method}{The method to use to compute the flow length (for now only 
the method "planar", using a single slope approximation, is implemented)}
}
\value{
The flow length numerical value.
}
\description{
Computes the flow length of a given matrix
}
\details{
This function computes the flow length of a given matrix, using a 
    planar approximation (the slope is constant across the whole matrix), 
    as per Rodriguez et al. (2017). 
  
  <Describe succintly (~100 wds) how this can be used as an indicator>
}
\examples{

fl_result <- flowlength_sews(arizona, slope = 20, cell_size = 1)

\dontrun{ 
# Compute the Z-score (standardized deviation to null distribution) and plot 
#   its variations along the gradient. This Z-score is suggested by 
#   Rodriguez et al. (2017) as an indicator of degradation. 
fl_test <- indictest(fl_result, nperm = 19)
plot(fl_test, what = "z_score")
}

}
\references{
Rodríguez, F., Á. G. Mayor, M. Rietkerk, and S. Bautista. 2017. A null model 
  for assessing the cover-independent role of bare soil connectivity as 
  indicator of dryland functioning and dynamics. Ecological Indicators.

<Mayor et al. 2013 would probably be a good reference here too>
}
