% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestdat.R
\docType{data}
\name{forestdat}
\alias{forestdat}
\title{A list of binary matrices and their associated parameters}
\format{A list of two components, matrices and parameters

  \enumerate{ 
    \item The `matrices` component contains ten binary matrices which are the 
      matrix states at the end of the simulations
    \item The `parameters` component contains a data frame with the value of 
      the parameters used to produce the simulations. 
  }}
\source{
Generated using the implementation of Kubo's model in caspr
}
\usage{
forestdat
}
\description{
A list of binary matrices and their associated parameters
}
\references{
Kubo, T., Iwasa, Y., & Furumoto, N. (1996). Forest spatial dynamics with gap
  expansion: Total gap area and gap size distribution. Journal of Theoretical
  Biology, 180(3), 229–246. http://doi.org/10.1006/jtbi.1996.0099
}
\keyword{datasets}

