% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestgap.R
\docType{data}
\name{forestgap}
\alias{forestgap}
\alias{forestgap.pars}
\title{A list of binary matrices and their associated parameters}
\format{A list of logical matrices which are the end results of simulations 
  from Kubo's Forest Gap model (see references).}
\source{
Generated using the implementation of Kubo's model in caspr 0.2.0 
  \url{https://github.com/fdschneider/caspr}.
}
\usage{
forestgap

forestgap.pars
}
\description{
A list of binary matrices and their associated parameters
}
\references{
Kubo, T., Iwasa, Y., & Furumoto, N. (1996). Forest spatial dynamics with gap
  expansion: Total gap area and gap size distribution. Journal of Theoretical
  Biology, 180(3), 229-246. \url{http://doi.org/10.1006/jtbi.1996.0099}
}
\keyword{datasets}
