% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation.R
\name{generic_spews}
\alias{generic_spews}
\title{(DEFUNCT) Generic Spatial Early-Warning signals}
\usage{
generic_spews(
  mat,
  subsize = 4,
  abs_skewness = FALSE,
  moranI_coarse_grain = FALSE
)
}
\arguments{
\item{mat}{A matrix (quantitative data), a binary matrix (TRUE/FALSE data), 
or a list of those}

\item{subsize}{The subsize used for the coarse-graining phase (see Details)}

\item{abs_skewness}{Should the absolute skewness be used instead of its 
raw values ?}

\item{moranI_coarse_grain}{Should the input matrix be coarse-grained before
computing the Moran's I indicator value ?}
}
\description{
Computation, significance assesment and display of spatial 
  generic early warning signals (Moran's I, variance and skewness)
}
\details{
\code{generic_spews} has been renamed, please use 
  \code{\link{generic_sews}} instead.
}
