% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_neighbors.R
\name{get_nb_coords}
\alias{get_nb_coords}
\title{Get the xy coords of neighboring cells}
\usage{
get_nb_coords(mat, X, nbmask, wrap)
}
\arguments{
\item{mat}{A 2D matrix}

\item{X}{A vector of coordinates (x,y) of the target cell}

\item{nbmask}{A neighboring "mask" that describes the neighbors to take into
              into account around the target cell (see examples).}

\item{wrap}{Whether to wrap or not when a point falls outside boundaries}
}
\value{
A two-columns matrix with the x,y coordinates of neighbors.
}
\description{

}
\examples{
nbmat4 <- matrix(c(0, 1, 0, # neighbor mask
                   1, 0, 1,
                   0, 1, 0), nrow=3, byrow=TRUE)
size <- 100
letmat <- matrix(sample(c(0,1,2), size, replace=TRUE), nrow=sqrt(size))

get_nb_coords(letmat, c(3,1), nbmat4, wrap = TRUE)

# Contrast with
get_nb_coords(letmat, c(3,1), nbmat4, wrap = FALSE)
}

