% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator_corrfunc.R
\name{indicator_corrfunc}
\alias{indicator_corrfunc}
\title{Spatial Warning Indicators: Correlation Function}
\usage{
indicator_corrfunc(mat)
}
\arguments{
\item{mat}{A binary matrix or a list of binary matrices.}
}
\value{
A numeric vector of correlation at different distances (from 0 to
matrix size/2 by increments of 1 unit) or a list of these vectors if input
is a list of binary matrices.
}
\description{
This function computes the correlation function of a binary
             matrix.
}
\details{
TODO: add details (reference ? what is a correlation function ?)
}

