% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_cumpsd.R
\name{indicator_cumpsd}
\alias{indicator_cumpsd}
\title{Get cumulative patch sizes.}
\usage{
indicator_cumpsd(x = NULL, patchvec = patchsizes(x))
}
\arguments{
\item{x}{A vector of patch sizes or a list of vectors.}

\item{patchvec}{a vector of patch sizes as produced by 
\code{\link{patchsizes}}}
}
\value{
A dataframe of the unique patch sizes \code{size}, the number of 
  patches of equal or larger size than each value in \code{size}, and a 
  probability that any given patch is equal or larger size than each value 
  in \code{size}.
}
\description{
Get cumulative patch sizes.
}
\examples{

data(B)
indicator_cumpsd(B)

}

