% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator_example.R
\name{indicator_example}
\alias{indicator_example}
\title{Spatial warning indicators: example}
\usage{
indicator_example(mat)
}
\arguments{
\item{mat}{A binary matrix or a list of binary matrices.}

\item{myparam1}{<Explanation>}

\item{myparam2}{<Explanation>}
}
\value{
An object of class <myclass>, or a list of objects of class <myclass>
        if `mat` was a list of binary matrices.
}
\description{
This indicator is a simple example. Please modify it ! :)
}
\references{
Thisguy et al., 2012 A new spatial indicator for critical points#'
}

