% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator_fitpsd.R
\name{indicator_fitpsd}
\alias{indicator_fitpsd}
\title{fit candidate cumulative patch size distribution functions.}
\usage{
indicator_fitpsd(x = NULL, cumpsd = indicator_cumpsd(x))
}
\arguments{
\item{x}{A patch size vector or a list of patch size vectors.}
}
\value{
A list object of class 'psdfit' containing the pooled cumulative patch size distribution data, as well as the AICs and model outputs of the candidate models.
}
\description{
fit candidate cumulative patch size distribution functions.
}
\details{
A S3 method for \code{summary()} does exist and only returns the model summary of the most parsimonious model.
}

