% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator_fractalgeometry.R
\name{indicator_fracgeo}
\alias{indicator_fracgeo}
\title{Spatial warning indicators: Fractal Geometry}
\usage{
indicator_fracgeo(mat = NULL, patchids = label(mat))
}
\arguments{
\item{mat}{A binary matrix or a list of binary matrices of length >=1.}

\item{patchids}{A integer matrix obtained from the function \code{\link{label}}.}
}
\value{
A list (of area and perimeter values for each patch in the input
  marix) if input is a single matrix (OR) a list of lists (of area and
  perimeter values for each patch) in every matrix of input list
}
\description{
function to calculate fractal geometries. Also calculates patch
  sizes (same as patch area).
}

