% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator_largestpatch.R
\name{indicator_largestpatch}
\alias{indicator_largestpatch}
\title{Spatial warning indicators: largest patch}
\usage{
indicator_largestpatch(mat = NULL, patchvec = patchsizes(mat))
}
\arguments{
\item{mat}{A binary matrix or a list of binary matrices.}

\item{patchvec}{A patch vector obtained from \code{\link{patchsizes}}.}
}
\value{
An object of class <myclass>, or a list of objects of class <myclass>
        if `mat` was a list of binary matrices.
}
\description{
Returns the size of the largest patch of the matrix.
}

