% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/morancorrelation_ews.R
\name{indicator_moran}
\alias{indicator_moran}
\title{Moran's I}
\usage{
indicator_moran(input)
}
\arguments{
\item{input}{A binary matrix or a list of binary matrix.}
}
\value{
An object of class numeric, or a list of objects of class numeric
        if `mat` was a list of binary matrices.
}
\description{
This function computes the Moran's I, i.e. a measure of the
             spatial autocorrelation.
}
\references{
P.A.P. Moran (1950) Notes on Continuous Stochastic Phenomena.
            Biometrika 37(1-2):17-23
}

