% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator_powerspectrum.R
\name{indicator_powerspectrum}
\alias{indicator_powerspectrum}
\title{Spatial warning indicators: Power Spectrum}
\usage{
indicator_powerspectrum(mat)
}
\arguments{
\item{mat}{A binary matrix or a list of binary matrices}
}
\value{
A list of lists(of theta and R spectrum values for distances 1 to
  SystemSize/2 for each matrix in the list) if input is a list (OR) a single
  list (of theta and R spectrum values for distances 1 to SystemSize/2) if
  input is a single matrix
}
\description{
Function to calculate theta and r spectrum.
}
\details{

}

