% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_sdr.R
\name{indicator_sdr}
\alias{indicator_sdr}
\title{Spectral Density Ratio (SDR) indicator}
\usage{
indicator_sdr(input, sdr_low_range = NULL, sdr_high_range = NULL,
  nreplicates = 999)
}
\arguments{
\item{input}{A matrix or a logical matrix (TRUE/FALSE), or a list of these.}

\item{sdr_low_range}{The range of values (in proportion) to 
use for the computation of the spectral density ratio.
For example, for the lowest 20% (default value), set sdr_low_range to 
c(0, .2).}

\item{sdr_high_range}{The range of values (in proportion) to 
use for the computation of the spectral density ratio. For example, for 
the higher 20% (default value), set sdr_high_range to 
c(.8, 1).}

\item{nreplicates}{The number of replicates to compute for the null 
distribution}
}
\value{
A list (or a list of lists if input was a list of matrices) with 
  components:
    \itemize{
      \item `value`: SDR of the matrix
    }
  If nreplicates is above 2, then the list has the following additional 
  components : 
    \itemize{
      \item `null_mean`: Mean SDR of the null distribution
      \item `null_sd`: SD of SDR in the null distribution
      \item `z_score`: Z-score of the observed value in the null distribution 
                         (value minus the null mean and divided by null 
                         standard deviation)
      \item `pval`: p-value based on the rank of the observed SDR
                      in the null distribution. A low p-value means that 
                      the indicator value is significantly higher than the 
                      null values. 
    }
}
\description{
Compute the ratio of low frequencies over high frequencies
  of the r-spectrum. It also computes a null value obtained by 
  randomizing the matrix.
}
\references{
Carpenter, S.R. & Brock, W.A. (2010). Early warnings of regime shifts in 
spatial dynamics using the discrete Fourier transform. Ecosphere, 1, art10
}
