% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator_skewness_main.R
\name{indicator_skewness_main}
\alias{indicator_skewness_main}
\title{Description: Spatial Early Warning Signals}
\usage{
indicator_skewness_main(rawmatrix, subsize = 2, detrending = FALSE,
  discrete = TRUE)
}
\arguments{
\item{rawmatrix}{is the input matrix.It should be a square matrix and contain either one or multiple snapshots of spatial data. If the input is a list of matrices,
the function should be called as lapply. Otherwise if the input is one matrix with one snapshot after the other in the next row, the function can be called in a regular way.}

\item{subsize}{is the dimension of the submatrix used to subsample the rawmatrix either to reduce the size of the original matrix}

\item{detrending}{logical. If TRUE data are detrended by removing the spatial mean. (Default is FALSE)}

\item{discrete}{logical. If TRUE the data represent discrete variables (like presence/absense), otherwise continuous data (like biomass density). Default is FALSE.}
}
\value{
\code{spatial_ews} returns a matrix that contains:

\item{stats}{the mean, skewness of both the original and the reduced data.}

\item{sk_null}{the skewness estimated on the null model based on the original data.}

\item{sk_null_red}{the skewness estimated on the reduced null model based on the subsampled data.}
}
\description{
\code{indicator_skewness_main} wrapper function used to estimate statistical properties from spatial data and compare results to null models
}

