% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator_variance_main.R
\name{indicator_variance_main}
\alias{indicator_variance_main}
\title{Description: Indicator variance}
\usage{
indicator_variance_main(rawmatrix, subsize = 2, detrending = FALSE,
  discrete = TRUE)
}
\arguments{
\item{rawmatrix}{is the input matrix.It should be a square matrix and contain either one or multiple snapshots of spatial data. If the input is a list of matrices,
the function should be called as lapply. Otherwise if the input is one matrix with one snapshot after the other in the next row, the function can be called in a regular way.}

\item{subsize}{is the dimension of the submatrix used to subsample the rawmatrix either to reduce the size of the original matrix}

\item{detrending}{logical. If TRUE data are detrended by removing the spatial mean. (Default is FALSE)}

\item{discrete}{logical. If TRUE the data represent discrete variables (like presence/absense), otherwise continuous data (like biomass density). Default is FALSE.}
}
\value{
\code{indicator_variance_main} returns a matrix that contains:

\item{stats}{the mean, variance of both the original and the reduced data.}

\item{var_null}{the variance estimated on the null model based on the original data.}

\item{var_null_red}{the variance estimated on the reduced null model based on the subsampled data.}
}
\description{
\code{indicator_variance_main} wrapper function used to estimate statistical properties from spatial data and compare results to null models
}

