% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generic_indic_indictest.R
\name{indictest.generic_spews_single}
\alias{indictest.generic_spews_single}
\title{Generic Spatial Early-Warnings}
\usage{
\method{indictest}{generic_spews_single}(obj, null_replicates = 999, ...)
}
\arguments{
\item{obj}{A \code{generic_spews} object as produced by 
\code{\link{generic_spews}}.}

\item{null_replicates}{The number of replicates to use to compute a null 
distribution}
}
\value{
An object of class \code{generic_spews_test} (actually, a 
  data.frame).
}
\description{
Assessment of the significance of generic early-warning signals
  values
}
\details{
The significance of generic early-warning signals can be estimated by 
  reshuffling the original matrix randomly. Indicators are then recomputed 
  on the shuffled matrices and the values obtained are used as a null 
  distribution. P-values are obtained based on the rank of the observered
  value in the null distribution.
}
\references{
Kéfi, S., Guttal, V., Brock, W.A., Carpenter, S.R., Ellison, A.M., 
  Livina, V.N., et al. (2014). Early Warning Signals of Ecological 
  Transitions: Methods for Spatial Patterns. PLoS ONE, 9, e92097.
}

