% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indictest.R
\name{indictest}
\alias{indictest}
\title{Significance-assessment of spatial early-warning signals}
\usage{
indictest(obj, null_replicates = 999, ...)
}
\arguments{
\item{obj}{A spatial warning object such as one produced by the *_spews 
function family}

\item{null_replicates}{How many times should the input matrix be randomized
to create the null distribution ?}
}
\value{
An object of class \code{*_spews_test} depending on the input object
  (actually, a data.frame)
}
\description{
Assess the significance of spatial early-warning values
}
\details{
This function implements the methods described in Kéfi et al. 
  (2014) that suggest randomizing the spatial matrix to compute a null 
  distribution. The observed values of indicators are tested against 
  this distribution to obtain p-values.
}
\references{
Kéfi, S., Guttal, V., Brock, W.A., Carpenter, S.R., Ellison, A.M., Livina, 
  V.N., et al. (2014). Early Warning Signals of Ecological Transitions: 
  Methods for Spatial Patterns. PLoS ONE, 9, e92097
}

