% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_kbdm.R
\name{kbdm_spews}
\alias{kbdm_spews}
\title{Indicator based on Kolmogorov Complexity}
\usage{
kbdm_spews(mat, subsize = 3)
}
\description{
Compute the Kolmogorov Complexity on a set of matrices, 
  using the Block Decomposition Method.
}
\details{
WriteMe
}
