% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patches.R
\name{label}
\alias{label}
\title{Labelling of unique patches.}
\usage{
label(mat)
}
\arguments{
\item{x}{A binary matrix or a list of binary matrices.}
}
\value{
A matrix containing ID numbers for each connected patch. Assuming
  4-cell neighborhood and periodic boundaries.
}
\description{
Labelling of unique patches.
}
\details{
The function is written in R and depends on the \code{mapping()}
  function of package caspr.
}

