% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/myfftshift_ews.R
\name{myfftshift}
\alias{myfftshift}
\title{myfftshift}
\usage{
myfftshift(X)
}
\arguments{
\item{X}{A binary square matrix.}
}
\value{
This function returns a matrix object. It is a shifted matrix.
}
\description{
This is R equivalent of fftshift function in matlab. This function is a dependency of rspec_ews.R and Indicator_SpectralFunction.R. It is called by rspec_ews.R.
}
\details{
Writen by Vishwesha Guttal, 7th Nov 2013.
}
\references{
Add Matlab function reference
}

