% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation.R
\name{patchdistr_spews}
\alias{patchdistr_spews}
\title{(DEFUNCT) Early-warning signals based on patch size distributions}
\usage{
patchdistr_spews(
  x,
  merge = FALSE,
  fit_lnorm = FALSE,
  best_by = "BIC",
  xmin = 1,
  xmin_bounds = NULL,
  wrap = FALSE
)
}
\arguments{
\item{x}{A logical matrix (TRUE/FALSE values) or a list of these}

\item{merge}{The default behavior is to produce indicators values for each 
matrix. If this parameter is set to TRUE then the patch size distributions 
are pooled together for fitting.}

\item{fit_lnorm}{When patch size distributions are compared, should we 
consider lognormal type ? (see details)}

\item{best_by}{The criterion to use to select the best fit (one of "AIC", 
"BIC" or "AICc")}

\item{xmin}{The xmin to be used to fit the patch size distributions. Use 
the special value "estimate" to compute first the xmin that produces 
the best power-law fit, then use this estimated value to fit all 
distributions.}

\item{xmin_bounds}{Bounds when estimating the xmin for power-law distributions}

\item{wrap}{Determines whether patches are considered to wrap around the 
matrix when reaching the side}
}
\description{
Compute early-warnings based on 
  patch size distributions and review/plot the results
}
\details{
\code{patchdistr_spews} has been renamed, please use 
  \code{\link{patchdistr_sews}} instead.
}
