% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_patch_indic.R
\name{patchdistr_spews}
\alias{patchdistr_spews}
\title{fit candidate cumulative patch size distribution functions.}
\usage{
patchdistr_spews(x)
}
\arguments{
\item{x}{A binary matrix or a list of binary matrices}
}
\value{
A list object of class 'psdfit' containing the pooled cumulative 
  patch size distribution data, as well as the AICs and model outputs of the 
  candidate models.
}
\details{
A S3 method for \code{summary()} does exist and only returns the 
  model summary of the most parsimonious model.
}
\examples{

data(B)
indicator_fitpsd(B)

}
\references{
Kefi, S., Rietkerk, M., Alados, C. L., Pueyo, Y., Papanastasis, 
V. P., ElAich, A., & De Ruiter, P. C. (2007). Spatial vegetation patterns 
and imminent desertification in Mediterranean arid ecosystems. 
Nature, 449(7159), 213-217.

Clauset, A., Shalizi, C. R., & Newman, M. E. (2009). 
Power-law distributions in empirical data. SIAM review, 51(4), 661-703.
}

