% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patches.R
\name{patches}
\alias{patches}
\title{Get patch sizes.}
\usage{
patches(x)
}
\arguments{
\item{mat}{A binary matrix or a list of binary matrices.}
}
\value{
A vector of patch sizes. An object of class 'patchvec', or a list of
 objects of class 'patchvec' if `mat` was a list of binary matrices.
}
\description{
G
}

