% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patchsizes.R
\name{patchsizes}
\alias{patchsizes}
\title{Get patch sizes.}
\usage{
patchsizes(x)
}
\arguments{
\item{x}{A binary matrix or a list of binary matrices.}
}
\value{
A vector of patch sizes. An object of class 'patchvec', or a list of
 objects of class 'patchvec' if `x` was a list of binary matrices.
}
\description{
Labels patches and counts patch size.
}

