% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generic_indic_methods.R
\name{plot.generic_spews}
\alias{plot.generic_spews}
\title{Generic spatial warning signals: plotting function}
\usage{
plot.generic_spews(obj, along = NULL)
}
\arguments{
\item{obj}{A \code{generic_spews} object (as provided by the 
\code{generic_spews} function).}

\item{along}{A vector providing values over which the indicator trend 
will be plotted.}
}
\description{
Plot generic early warning signals trends
}
\details{
A ggplot object (usually displayed immediately when plot() is 
  called interactively).

Since this function returns a ggplot object, it can be later 
  modified to add other graphical elements (e.g. axis names or annotations).
}
\seealso{
\code{\link{generic_spews}}, 
  \code{\link{plot.generic_spews_test}}

 @examples
 data(forestdat)
 
 genindic <- generic_spews(forestdat[['matrices']])
 plot(genindic, along = forestdat[['parameters']][ ,'delta']) 
 
 # The plot can be modified using ggplot functions
 if ( require(ggplot2) ) { 
   plot(genindic, along = forestdat[['parameters']][ ,'delta']) + 
     xlab('delta') + 
     theme_minimal() 
 }
}

