% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generic_indic.R
\name{plot.generic_spews_summary}
\alias{plot.generic_spews_summary}
\title{Plots the results of a generic spatial warning signal summary}
\usage{
\method{plot}{generic_spews_summary}(obj, along = unique(obj[, "replicate"]),
  what = "value", display_null = TRUE)
}
\arguments{
\item{obj}{A \code{generic_spews_summary} object (as provided by the 
\code{summary} methods for \code{generic_spews} objects).}

\item{along}{A vector providing values over which the indicator trend 
will be plotted.}

\item{what}{The variable to plot. The default is to plot the value of the 
indicator, but other variables can be chosen (e.g. p-value).}

\item{display_null}{Controls the display of a grey ribbon representing the 
95% and 5% quantile of the null distribution}
}
\value{
A ggplot object (usually displayed immediatelly when called at the 
  prompt).
}
\description{
Plotting functions for generic early warning signals
}
\details{
Since this function returns a ggplot object, it can be later 
  modified to add other graphical elements (e.g. axis names or annotations).
}
\examples{
data(B)
result <- generic_spews( as.binary_matrix(B) )
plot(summary(result))

}
\seealso{
\code{\link{generic_spews}}, \code{\link{summary.generic_spews}}
}

