% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generic_indic_indictest_methods.R
\name{plot.generic_spews_test}
\alias{plot.generic_spews_test}
\title{Generic spatial warning signals: plotting function}
\usage{
plot.generic_spews_test(obj, along = NULL, what = "value",
  display_null = TRUE, ...)
}
\arguments{
\item{obj}{A \code{generic_spews} object (as provided by the 
\code{generic_spews} function).}

\item{along}{A vector providing values over which the indicator trend 
will be plotted.}

\item{what}{The trendline to be displayed. Defaults to the indicator's 
values ("value") but other metrics can be displayed. Correct values are 
"value", "pval" or "z_score".}

\item{display_null}{Chooses whether a grey ribbon should be added to reflect
the null distribution. Note that it can not be displayed when the trend 
line reflects something else than the indicator values (when \code{what} 
is not set to "value").}

\item{...}{Ignored.}
}
\description{
Plot generic early warning signals trends (with null 
  distributions)
}
\details{
A ggplot object (usually displayed immediately when plot() is 
  called interactively).

Since this function returns a ggplot object, it can be later 
  modified to add other graphical elements (e.g. axis names or annotations).
}
\seealso{
\code{\link{generic_spews}}, 
  \code{\link{plot.generic_spews_test}}

 @examples
 data(forestdat)
 
 genindic <- indictest(generic_spews(forestdat[['matrices']]))
 plot(genindic, along = forestdat[['parameters']][ ,'delta']) 
 
 # The plot can be modified using ggplot functions
 if ( require(ggplot2) ) { 
   plot(genindic, along = forestdat[['parameters']][ ,'delta']) + 
     xlab('delta') + 
     theme_minimal() 
 }
}

