% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_cumpsd.R
\name{plot.indicator_cumpsd}
\alias{plot.indicator_cumpsd}
\title{Plot a patch-size cumulative distribution}
\usage{
\method{plot}{indicator_cumpsd}(x, x.log = TRUE, y.log = TRUE,
  add.line = TRUE, facets = FALSE)
}
\arguments{
\item{x}{An `indicator_cumpsd` object as produced by `indicator_cumpsd`}

\item{x.log}{Whether to use log scale on x axis}

\item{y.log}{Whether to use log scale on y axis}

\item{add.line}{Adds a straight regression line to the plot}

\item{facets}{Produce one plot facet per replicate or merge them altogether
in one graph.}
}
\examples{

# An example with a list
data(L)
result <- indicator_cumpsd(L)
plot(result) 
plot(result, facets = TRUE)

}

