% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_fitpsd.R
\name{plot.psdfit}
\alias{plot.psdfit}
\title{Plot a `psdfit` object}
\usage{
\method{plot}{psdfit}(x, all.models = FALSE)
}
\arguments{
\item{x}{A `psdfit` object as returned by `indicator_fitpsd`}

\item{all.models}{Should all fitted models be displayed on the plot ?}
}
\examples{
data(B)
fit <- indicator_fitpsd(B)
plot(B)

}

