% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_indic_indictest.R
\name{plot.simple_sews_test}
\alias{plot.simple_sews_test}
\title{Spatial early-warning signals: display of trends}
\usage{
\method{plot}{simple_sews_test}(x, along = NULL, what = "value",
  display_null = TRUE, ...)
}
\arguments{
\item{what}{The trendline to be displayed. Defaults to the indicator's 
values ("value") but other metrics can be displayed. Correct values are 
"value", "pval" or "z_score".}

\item{display_null}{Chooses whether a grey ribbon should be added to reflect
the null distribution. Note that it can not be displayed when the trend 
line reflects something else than the indicator values (when \code{what} 
is not set to "value").}
}
\description{
Spatial early-warning signals: display of trends
}
