% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_spectral_indic_plot.R
\name{plot.spectral_spews_test}
\alias{plot.spectral_spews_test}
\title{Spectral spatial early-warning signals: SDR plotting function}
\usage{
plot.spectral_spews_test(obj, along = NULL, what = "value",
  display_null = TRUE)
}
\arguments{
\item{obj}{An \code{spectral_spews_test} object as produced by \link{indictest}}

\item{along}{An optional vector of values along which the trend is to be 
displayed. The length of the vector must be equal to the number of 
indicator values provided in the object \code{obj}.}

\item{what}{The trendline to be displayed. Defaults to the indicator's 
values ("value") but other metrics can be displayed. Correct values are 
"value", "pval" or "z_score".}

\item{display_null}{Chooses whether a grey ribbon should be added to reflect
the null distribution. Note that it can not be displayed when the trend 
line reflects something else than the indicator values (when \code{what} 
is not set to "value").}
}
\value{
A ggplot object (usually displayed immediately when plot() is 
  called interactively).
}
\description{
This method reads the Spectral Density Ratio (SDR) values 
  produced by \code{\link{indictest}} and plots the trend with null values.
}
\details{
Since this function returns a ggplot object, it can be later 
  modified to add other graphical elements (e.g. axis names or annotations).
}
\examples{

data(forestdat)
sdr_test <- indictest( spectral_spews(forestdat[['matrices']]) )
plot(sdr_test)

# Display the z_score along the vector of parameter values
plot(sdr_test, 
     along = forestdat[['parameters']][ ,'delta'], 
     what = 'z_score')

if ( require(ggplot2) ) { 
  plot(sdr_test, along = forestdat[['parameters']][ ,'delta']) + 
    xlab('delta') +  
    theme_minimal() 
}
}

