% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_methods.R
\name{print.spindic}
\alias{print.spindic}
\title{Print a spatial indicator object}
\usage{
\method{print}{spindic}(obj, verbose = FALSE)
}
\arguments{
\item{obj}{An `spindic` object}

\item{verbose}{If `TRUE` then print all indicator objects using 
`print.default`}
}
\description{
Print method for `spindic` objects (spatial 
  indicators).
}

