% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_skewness.R
\name{raw_cg_skewness}
\alias{raw_cg_skewness}
\title{Skewness of a coarse-grained matrix}
\usage{
raw_cg_skewness(mat, subsize, absolute)
}
\arguments{
\item{mat}{A matrix that can contain logical (with TRUE/FALSE values)}

\item{subsize}{The size of the submatrices used for coarse-graining}

\item{absolute}{Whether to return the unmodified skewness, or its 
absolute value}
}
\description{
Compute the skewness of a given matrix after coarse-graining
}
