% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_flowlength.R
\name{raw_flowlength_planar}
\alias{raw_flowlength_planar}
\title{Flow length (planar approximation)}
\usage{
raw_flowlength_planar(mat, slope, cell_size)
}
\arguments{
\item{mat}{The input matrix (must be a logical matrix)}

\item{slope}{The slope of the area documented by the matrix (in degree).}

\item{cell_size}{The horizontal size of a cell in the matrix (as viewed 
from above).}
}
\value{
The flow length numerical value.
}
\description{
Compute a simple approximation of the flow length assuming a 
  constant slope
}
\details{
This function computes the flow length of a given matrix, using a 
    planar approximation (the slope is constant across the whole matrix), 
    as per Rodriguez et al. (2017). See \code{\link{flowlength_sews}} for 
    more details.
}
\examples{

raw_flowlength_planar(arizona[[1]], slope = 20, cell_size = 1)




}
\references{
Rodríguez, F., Á. G. Mayor, M. Rietkerk, and S. Bautista. 2017. A null model 
  for assessing the cover-independent role of bare soil connectivity as 
  indicator of dryland functioning and dynamics. Ecological Indicators.

<Mayor et al. 2013 ?>
}
\seealso{
\code{\link{flowlength_sews}}
}
