% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_variogram_indic.R
\name{raw_structvar}
\alias{raw_structvar}
\title{Structural variance}
\usage{
raw_structvar(mat, model = "sph", nmax = 100000L, nbins = 32, cutoff = NULL)
}
\arguments{
\item{mat}{A matrix}

\item{model}{The variogram model to use, either "sph" (for a spherical
model) or "exp" (for an exponential model)}

\item{nmax}{The maximum number of pairs of cells to use when computing the 
variogram}

\item{nbins}{Number of distance bins to use to compute the variogram}

\item{cutoff}{Maximum distance to consider in the variogram. If NULL, then 
a distance equal to one third of the diagonal of the matrix is used}
}
\description{
Compute the structural variance on a matrix.
}
\examples{

raw_structvar(serengeti[[5]])

}
\seealso{
raw_variogram_metrics, variogram_sews
}
