% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_variogram_indic.R
\name{raw_variogram_metrics}
\alias{raw_variogram_metrics}
\title{Variogram parameters}
\usage{
raw_variogram_metrics(mat, subset_frac = 0.5, model = "sph")
}
\arguments{
\item{mat}{A matrix (TRUE/FALSE values) or a list of matrices}

\item{subset_frac}{The number of points to consider in the matrix (as a 
proportion of the total number of cells) used to computed the variogram. 
A lower value here will speed up computations, at the expense of precision.}

\item{model}{The variogram model to use, either "sph" (for a spherical model)
or "exp" (for an exponential model)}
}
\description{
Compute the nugget, partial sill, correlation range and 
  structural variance on a matrix.
}
