% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-tags.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{moment}
\alias{plan}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{future}{\code{\link[future]{plan}}}

  \item{moments}{\code{\link[moments]{moment}}}
}}

