% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rspectrum}
\alias{rspectrum}
\title{r-spectrum}
\usage{
rspectrum(mat)
}
\arguments{
\item{mat}{A matrix of logical or numeric values}
}
\value{
A data.frame with two columns: \code{dist}, the wave number and 
  \code{rspec}, the normalized value of the r-spectrum
}
\description{
Compute the r-spectrum of a matrix
}
\seealso{
\code{\link{spectral_spews}}, \code{\link{indicator_sdr}}
}
