% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serengeti.R
\docType{data}
\name{serengeti}
\alias{serengeti}
\alias{serengeti.rain}
\title{Serengeti dataset}
\format{A list of logical matrices}
\source{
Extraceted from Eby's et al (2017) supplementary material 
  \url{https://github.com/tee-lab/spacetime-csd/}
}
\usage{
serengeti

serengeti.rain
}
\description{
A set of aerial photography taken along a rainfall gradient 
  in Tanzania.
}
\references{
Eby, S., Agrawal, A., Majumder, S., Dobson, A.P. & Guttal, V. (2017). 
Alternative stable states and spatial indicators of critical slowing down 
along a spatial gradient in a savanna ecosystem: Eby et al. Global Ecology 
and Biogeography, 26, 638-649
}
\keyword{datasets}
