% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_indic_indictest.R,
%   R/simple_indic_methods.R
\name{plot.simple_sews_test}
\alias{plot.simple_sews_test}
\alias{plot.simple_sews_list}
\title{Spatial early-warning signals: display of trends}
\usage{
\method{plot}{simple_sews_test}(x, along = NULL, what = "value", display_null = TRUE, ...)

\method{plot}{simple_sews_list}(x, along = NULL, ...)
}
\arguments{
\item{x}{A \code{simple_sews_test} object (as provided by **_sews functions, such 
as \code{generic_sews()}}

\item{along}{A vector providing values over which the indicator trend 
will be plotted. If \code{NULL} then the values are plotted sequentially 
in their original order.}

\item{what}{The trendline to be displayed. Defaults to the indicator's 
values ("value") but other metrics can be displayed. Accepted values are 
"value", "pval" or "z_score".}

\item{display_null}{Chooses whether a grey ribbon should be added to reflect
the null distribution. Note that it can not be displayed when the trend 
line reflects something else than the indicator values (when \code{what} 
is not set to "value").}

\item{...}{Other arguments are ignored.}
}
\description{
Spatial early-warning signals: display of trends
}
\details{
Note that the produced plot is adjusted depending on whether 
  \code{along} is numeric or not.
}
