% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial_ews.R
\name{spatial_ews}
\alias{spatial_ews}
\title{Description: Spatial Early Warning Signals}
\usage{
spatial_ews(rawmatrix, numsnaps = 1, discrete = FALSE, subsize = 5,
  nullmodel = 0, iter = 50, detrending = FALSE, logtransform = FALSE)
}
\arguments{
\item{rawmatrix}{is the input matrix.It should be a square matrix and contain either one or multiple snapshots of spatial data.}

\item{numsnaps}{is the number of snapshots in the rawmatrix (DEFAULT = numrow/numcols)}

\item{discrete}{logical. If TRUE the data represent discrete variables (like presence/absense), otherwise continuous data (like biomass density). Default is FALSE.}

\item{subsize}{is the dimension of the submatrix used to subsample the rawmatrix either to reduce the size of the original matrix to  (DEFAULT value = 5)}

\item{nullmodel}{generates a null distribution of spatial statistics. Value 0 means the null model is a reshuffled matrix of original data. Value of 1 means a random matrix with Gaussian distribution of same mean and variance as original data. (DEFAULT is 0)}

\item{iter}{is the number of iterations that the null model runs. (DEFAULT is 50)}

\item{detrending}{logical. If TRUE data are detrended by removing the spatial mean. (Default is FALSE)}

\item{logtransform}{logical. If TRUE data are logtransformed prior to analysis as log(X+1). (Default is FALSE)}
}
\value{
\code{spatial_ews} returns a matrix that contains:

\item{stats}{the mean, std deviation, skewness and Moran correlation at lag 1 of both the original and the subsampled data.}

\item{thetaspectr}{the theta spectrum of the original data.}

\item{rhospectr}{the rho spectrum of the original data.}

\item{SD_null}{the std deviatins estimated on the null model based on the original data.}

\item{SK_null}{the skewness estimated on the null model based on the original data.}

\item{CORR_null}{the Moran correlation at lag 1 estimated on the null model based on the original data.}

\item{SD_null_red}{the std deviations estimated on the null model based on the subsampled data.}

\item{SK_null_red}{the skewness estimated on the null model based on the subsampled data.}

\item{CORR_null_red}{the Moran correlation at lag 1 estimated on the null model based on the subsampled data.}

In addition, \code{spatial_ews} returns for each snapshot a plot with the original data, the 2dim DFT on the original data, the theta and rho spectra, the patch size distribution (if the data are discrete), and a table of the spatial statistics compared to the 5 and 95 percentiles estimated from the chosen null model.
If the input rawmatrix is of multiple snapshots, two extra plots are returned for visualising the trends of the spatial indicators for the original and reduced data.
}
\description{
\code{spatial_ews} wrapper function used to estimate statistical properties from spatial data and compare results to null models
}
\details{
see ref below
}
\examples{
data(desertification_BW)
out=spatial_ews(desertification_BW, numsnaps=1, discrete=FALSE, subsize=5, nullmodel=0, iter = 50, detrending = FALSE, logtransform = FALSE)
}
\author{
Vishewesha Guttal, Sonia Kefi, Vasilis Dakos \email{vasilis.dakos@gmail.com}
}
\references{
Kefi S. et al (2013). Early warning signals of ecological transitions: Methods for spatial patterns. PLoS One (in review)
}
\seealso{
\code{\link{generic_ews}}  \code{\link{ddjnonparam_ews}}  \code{\link{bdstest_ews}}  \code{\link{sensitivity_ews}}  \code{\link{surrogates_ews}}  \code{\link{ch_ews}}  \code{\link{movpotential_ews}}  \code{\link{livpotential_ews}}
}
\keyword{early-warning}

