% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialwarnings-deprecated.R
\name{spatialwarnings-deprecated}
\alias{spatialwarnings-deprecated}
\title{Deprecated functions in package \pkg{spatialwarnings}.}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned.
}
\section{\code{generic_spews}}{

  For \code{\link{generic_spews}}, use \code{\link{generic_sews}}.
}

\section{\code{spectral_spews}}{

  For \code{\link{spectral_spews}}, use \code{\link{spectral_sews}}.
}

\section{\code{patchdistr_spews}}{

  For \code{\link{patchdistr_spews}}, use \code{\link{patchdistr_sews}}.
}

\keyword{internal}
