% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialwarnings.R
\docType{package}
\name{spatialwarnings}
\alias{spatialwarnings}
\alias{spatialwarnings-package}
\alias{spatialwarnings-package}
\title{Early Spatial-Warnings of Ecosystem Degradation}
\section{can split the help page into several sections to show}{
NA
}

