% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation.R
\name{spectral_spews}
\alias{spectral_spews}
\title{(DEFUNCT) Spectrum-based spatial early-warning signals.}
\usage{
spectral_spews(mat, sdr_low_range = NULL, sdr_high_range = NULL, quiet = FALSE)
}
\arguments{
\item{mat}{The input matrix or a list of matrices.}

\item{sdr_low_range}{The range of values (in proportion) to 
use for the computation of the spectral density ratio.
For example, for the lowest 20% (default value), set sdr_low_range to 
c(0, .2). See also the Details section.}

\item{sdr_high_range}{The range of values (in proportion) to 
use for the computation of the spectral density ratio. For example, for 
the higher 20% (default value), set sdr_high_range to 
c(.8, 1). See also the Details section.}

\item{quiet}{Do not display some warnings}
}
\description{
Computation of spatial early warning signals based on spectral
  properties.
}
\details{
\code{spectral_spews} has been renamed, please use 
  \code{\link{spectral_sews}} instead.
}
