% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_spectral_indic.R
\name{spectral_spews}
\alias{spectral_spews}
\title{Spectrum-based spatial early-warning signals.}
\usage{
spectral_spews(input, sdr_low_range = NULL, sdr_high_range = NULL)
}
\arguments{
\item{input}{The input matrix or a list of matrices.}

\item{sdr_low_range, }{sdr_high_range The range of values (in proportion) to 
use for the computation of the spectral density ratio (default: lower and 
higher 20%). For example, for the lowest 20%, set sdr_low_range to 
c(0, .2). See also the Details section.}
}
\value{
An object of class \code{spectral_spews_list} or 
  \code{spectral_spews_single} depending on whether the input was a list of 
  matrices or a single matrix.
}
\description{
Computation of spatial early warning signals based on spectral
  properties.
}
\details{
Spectral early warning signals are based on the fact that some dynamical 
    systems can exhibit an increase in their spatial structure when 
    approaching a transition (e.g. by forming patches or scale-free patterns).
    This is expected to be reflected in the spectrum of the spatial structure
    by an increase in lower frequencies ("reddening" of the spectrum). 
    
  This functions allows computing the radial-spectrum which gives the relative 
    dominance of each space-scale over a range of distances, from 1 to 
    \code{N/2} (N being the minimum between the number of rows and columns). 
    If the matrix is not square, then it is cropped to biggest square that 
    fits on the left side of the matrix. 
  
  Additionnaly, it summarizes this spectrum into a Spectral 
    Density Ratio (SDR), which is the ratio of low frequencies over 
    high frequencies of the r-spectrum.
}
\examples{

data(forestdat) 

spec_indic <- spectral_spews(forestdat[['matrices']], 
                             sdr_low_range  = c(0, .2), 
                             sdr_high_range = c(.2, 1))

summary(spec_indic)

# Display SDR trend
plot(spec_indic)

# Display spectra
plot_spectrum(spec_indic)

}
\references{
Kéfi, S., Guttal, V., Brock, W.A., Carpenter, S.R., Ellison, A.M., 
  Livina, V.N., et al. (2014). Early Warning Signals of Ecological 
  Transitions: Methods for Spatial Patterns. PLoS ONE, 9, e92097.
}

