% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_generic_indic.R
\name{summary.generic_spews}
\alias{summary.generic_spews}
\title{Generic spatial warning signals: summary function}
\usage{
\method{summary}{generic_spews}(obj, null_replicates = 999)
}
\arguments{
\item{obj}{A \code{generic_spews} object (as provided by the
  \code{generic_spews} function).}

\item{null_replicates}{The number of replicates to generate in the null
  distribution to test significance}
}
\description{
Summary and significance testing function for generic
  spatial early-warning signals.
}
\details{
This function returns internally a \code{data.frame} of five columns
  containing :
  \enumerate {
    \item \code{value} The indicator value
    \item \code{null_mean} The mean of the null distribution
    \item \code{null_sd} The standard deviation of the null distribution
    \item \code{null_95} The 95%th quantile
    \item \code{null_05} The 5%th quantile
    \item \code{z_score} The z_score of the observed value in the null distribution
    \item \code{pval} The p-value (based on the rank of the observed value in the null
            distribution)
  }
}
\examples{
data(B)
result <- generic_spews( as.binary_matrix(B) )
summary(result)
}
\seealso{
\code{\link{plot.generic_spews_summary}}
}

