% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarse_graining.R
\name{coarse_grain}
\alias{coarse_grain}
\title{Matrix coarse-graining}
\usage{
coarse_grain(mat, subsize)
}
\arguments{
\item{mat}{TODO}

\item{subsize}{TODO}
}
\value{
TODO
}
\description{
Reduce a matrix size by coarse-graining
}
