% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_generic_indic.R
\name{generic_spews}
\alias{generic_spews}
\title{Generic spatial warning signals}
\usage{
generic_spews(mat, discrete = is.binary_matrix(mat), subsize = 5,
  detrend = FALSE)
}
\arguments{
\item{mat}{A matrix (quantitative data), a binary matrix (qualitative data),
  or a list of those}

\item{discrete}{Whether the passed data is qualitative or quantitative. If
  the passed data is a binary matrix, then the parameter is set to TRUE.}

\item{subsize}{The subsize for the coarse-graining in case the passed matrix
  is qualitative}

\item{detrend}{Whether to substract the mean or not to the input matrix
  (detrend)}
}
\value{
An object of class \code{generic_spews_single} if the passed object
  was a single matrix or an object of class \code{generic_spews_list} if
  it was a list.
}
\description{
Computation of spatial generic early warning signals (Moran's I,
  variance and skewness)
}
\details{
Before a critical point, a spatial dynamical system is expected to
  show an increase in spatial correlation at lag-1 (measured by Moran's I),
  in variance and in skewness. These functions provides a workflow to
  compute and test those indicators (see
  \code{\link{indicator_moran}}, \code{\link{indicator_variance}} and
  \code{\link{indicator_skewness}} details about null models).
}
\references{
Dakos, V., van Nes, E. H., Donangelo, R., Fort, H., &
  Scheffer, M. (2010). Spatial correlation as leading indicator of
  catastrophic shifts. Theoretical Ecology, 3(3), 163-174.

  Guttal, V., & Jayaprakash, C. (2008). Spatial variance and spatial
  skewness: leading indicators of regime shifts in spatial ecological
  systems. Theoretical Ecology, 2(1), 3–12.
}
\seealso{
\code{\link{indicator_moran}}, \code{\link{indicator_variance}} and
  \code{\link{indicator_skewness}}
}

