% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitpsd.R, R/indicator_cumpsd.R
\name{indicator_cumpsd}
\alias{indicator_cumpsd}
\title{Get cumulative patch sizes.}
\usage{
indicator_cumpsd(x = NULL, patchvec = patchsizes(x))

indicator_cumpsd(x = NULL, patchvec = patchsizes(x))
}
\arguments{
\item{x}{A vector of patch sizes or a list of vectors.}

\item{x}{A vector of patch sizes or a list of vectors.}
}
\value{
A dataframe of the unique patch sizes \code{size}, the number of patches of equal or larger size than each value in \code{size}, and a probability that any given patch is equal or larger size than each value in \code{size}.

A dataframe of the unique patch sizes \code{size}, the number of patches of equal or larger size than each value in \code{size}, and a probability that any given patch is equal or larger size than each value in \code{size}.
}
\description{
Get cumulative patch sizes.

Get cumulative patch sizes.
}

