% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator_moran.R
\name{indicator_moran}
\alias{indicator_moran}
\title{Moran's Index at lag of 1}
\usage{
indicator_moran(input, subsize = 2, detrending = FALSE, discrete = TRUE)
}
\arguments{
\item{input}{An matrix or a list of matrix object. It should
  be a square matrix}

\item{subsize}{logical. Dimension of the submatrix used to coarse-grain the
  original matrix.}

\item{detrending}{If TRUE data are detrended by removing the spatial mean.
  (Default is FALSE)}

\item{discrete}{logical. If TRUE the data represent discrete variables (like
  presence/absense), otherwise continuous data (like biomass density).
  Defaults to FALSE.}
}
\value{
A list (or a list of list if input was a list of matrix
object) of:
  \itemize{
    \item `Mean`: Landscape mean cover
    \item `MoranNullCorr`: Autocorrelation computed in the randomized matrix.
    \item `MoranCorr`: Spatial autocorrelation of the matrix
  }
}
\description{
This functions computes the Moran's spatial correlation index
  (with lag one). It also computes a null value obtained by randomizing
  the matrix.
}

