% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_plrange.R
\name{indicator_plrange}
\alias{indicator_plrange}
\title{TODO}
\usage{
indicator_plrange(mat, merge = FALSE, xmin_bounds = NULL)
}
\arguments{
\item{mat}{TODO}

\item{merge}{TODO}

\item{xmin_bounds}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO

}
\references{
TODO
}
