% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_psdtype.R
\name{indicator_psdtype}
\alias{indicator_psdtype}
\title{Change in patch-size distributions types}
\usage{
indicator_psdtype(x, xmin = 1, merge = FALSE, fit_lnorm = FALSE,
  xmin_bounds = NULL, best_by = "AIC", wrap = FALSE)
}
\arguments{
\item{x}{A logical (TRUE/FALSE values) matrix or a list of these.}

\item{xmin}{The xmin to be used to fit the patch size distributions. Use 
the special values "estimate" to use an estimated xmin for each fit}

\item{merge}{If input is a list, then merge all the observed patch-size
distributions in a single one to obtain a better fit.}

\item{fit_lnorm}{Fit also a log-normal distribution}

\item{xmin_bounds}{Restrict the possible xmins in this range (defaults to 
the whole range of observed patch sizes)}

\item{best_by}{The criterion used to select the best distribution type 
(one of \code{"AIC"}, \code{"BIC"} or \code{"AICc"}).}

\item{wrap}{Detemines whether patches are considered to wraparound when 
reaching one side of the matrix.}
}
\value{
A data.frame (or a list of these if x is a list) with the 
  following columns:
    \itemize{
      \item `method` the method used for fitting (currently: only 
         log-likelihood is implemented, "ll")
      \item `type` the type of distribution fit
      \item `npars` the number of parameters of the distribution type
      \item `AIC`, `AICc` and `BIC` the values for Akaike Information 
        Criterion (or the corrected for small samples equivalent AICc), 
        and Bayesion Information Criterion (BIC)
      \item `best` A logical vector indicating which distribution is the 
        best fit 
      \item `expo`, `rate`, `meanlog`, `sdlog` the estimates for distribution
        parameters. 
      \item 'percolation' A logical value indicating whether there is 
        \code{\link{percolation}} in the system. 
    }
}
\description{
This functions fits different patch size distributions 
 types (power-law, longnormal, exponential and truncated power-law) to 
 the patches contained in a matrix. The distributions are returned with 
 their corresponding AIC, BIC and AICc to select the best fit.
}
\examples{

data(forestgap)

indicator_psdtype(forestgap)


}
\references{
Kefi, S., Rietkerk, M., Roy, M., Franc, A., de Ruiter, P.C. & Pascual, M.
(2011). Robust scaling in ecosystems and the meltdown of patch size
distributions before extinction: Patch size distributions towards 
extinction. Ecology Letters, 14, 29-35.

Kéfi, S., Rietkerk, M., Alados, C.L., Pueyo, Y., Papanastasis, V.P., ElAich,
A., et al. (2007). Spatial vegetation patterns and imminent desertification
in Mediterranean arid ecosystems. Nature, 449, 213-217.

Clauset et al. #TODO
}
