% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_sdr.R
\name{indicator_sdr}
\alias{indicator_sdr}
\title{Spectral Density Ratio (SDR) indicator}
\usage{
indicator_sdr(input, sdr_low_range = NULL, sdr_high_range = NULL,
  nreplicates = 999)
}
\arguments{
\item{input}{A matrix or a logical matrix (TRUE/FALSE), or a list of these.}

\item{sdr_low_range}{The range of values (in proportion) to 
use for the computation of the spectral density ratio.
For example, for the lowest 20% (default value), set sdr_low_range to 
c(0, .2).}

\item{sdr_high_range}{The range of values (in proportion) to 
use for the computation of the spectral density ratio. For example, for 
the higher 20% (default value), set sdr_high_range to 
c(.8, 1).}

\item{nreplicates}{The number of replicates to compute for the null 
distribution}

\item{quiet}{Disable some warnings}
}
\value{
A single value containing the SDR value
}
\description{
Compute the ratio of low frequencies over high frequencies
  of the r-spectrum. It also computes a null value obtained by 
  randomizing the matrix.
}
\references{
? Biggs et al. 2008 ? (Vishu suggestion IIRC [Alex])
}

