% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_sdr.R
\name{indicator_sdr}
\alias{indicator_sdr}
\title{Spectral Density Ratio (SDR) indicator}
\usage{
indicator_sdr(input, low_range, high_range, nreplicates = 499)
}
\arguments{
\item{low_range}{A numeric vector of the form \code{c(min, max)} describing the 
range of values considered as low frequencies}

\item{high_range}{A numeric vector of the form \code{c(min, max)} describing the 
range of values considered as high frequencies}

\item{nreplicates}{The number of replicates to compute for the null 
distribution}

\item{mat}{A matrix or a binary_matrix, or a list of these.}
}
\value{
A single value containing the SDR value
}
\description{
Compute the ratio of low frequencies over high frequencies
  of the r-spectrum. It also computes a null value obtained by 
  randomizing the matrix.
}
\references{
? Biggs et al. 2008 ? (Vishu suggestion IIRC [Alex])
}

