% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_skewness.R
\name{indicator_skewness}
\alias{indicator_skewness}
\title{Skewness indicator}
\usage{
indicator_skewness(input, subsize = 2, detrending = FALSE,
  nreplicates = 499)
}
\arguments{
\item{input}{A square binary matrix or a list of square binary matrices.}

\item{subsize}{logical. Dimension of the submatrix used to coarse-grain the 
original matrix.}

\item{detrending}{If TRUE data are detrended by removing the spatial mean. 
(Default is FALSE).}

\item{nreplicates}{Number of replicates to produce to estimate null 
distribution of index.}
}
\value{
A list (or a list of list if input was a list of matrices) with 
  components:
    \itemize{
      \item `mean`: Landscape mean cover
      \item `value`: Spatial skewness of the matrix
    }
  If nreplicates was above 2, then the list has the following additional 
  components : 
    \itemize{
      \item `null_mean`: Mean skewness of the null distribution
      \item `null_sd`: SD of skewness in the null distribution
      \item `z_score`: Z-score of the observed value in the null distribution
      \item `pval`: p-value based on the rank of the observed skewness
                      in the null distribution.
    }
}
\description{
This functions computes the skewness critical point indicator. 
  It also computes a null value obtained by randomizing 
  the matrix.
}
\examples{
data(B)
indicator_skewness(B)


}
\references{
Guttal, V., and Jayaprakash, C. (2009). Spatial variance and 
spatial skewness: leading indicators of regime shifts in spatial 
ecological systems. Theoretical Ecology, 2(1), 3-12.
}

