% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Labelling of patches.}
\usage{
label(mat)
}
\arguments{
\item{x}{A binary matrix or a list of binary matrices.}
}
\value{
A matrix containing ID numbers for each connected patch, assuming
  4-cell neighborhood and periodic boundaries.
}
\description{
Labelling of patches.
}
\details{
The function is written in R and depends on the \code{mapping()}
  function of package caspr.
}
\examples{
data(B)
  par(mar=c(0,0,0,0))
  image(B, xaxt = "n", yaxt = "n", asp = 1, bty = "n", col = c("white", "black"))
  M <- label(B)
  image(M, xaxt = "n", yaxt = "n", asp = 1, bty = "n",
     col = rep(rainbow(24), 100) )
}

