% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Labelling of unique patches.}
\usage{
label(mat, nbmask = matrix(c(0, 1, 0, 1, 0, 1, 0, 1, 0), ncol = 3),
  wrap = TRUE)
}
\arguments{
\item{mat}{A binary matrix}

\item{nbmask}{a "neighboring mask": a matrix with odd dimensions describing
       which neighbors are to be considered around a cell (see examples).}

\item{wrap}{Whether to wrap around lattice boundaries (`TRUE`/`FALSE`).}
}
\value{
A matrix containing ID numbers for each connected patch. Assuming
  4-cell neighborhood and periodic boundaries.
}
\description{
Labelling of unique patches.
}
\examples{
data(B)
image(label(B))


# With 8-way neighborhood mask
nbmask8 <- matrix(c(1,1,1,
                    1,0,1,
                    1,1,1), ncol=3)
image(label(B, nbmask8, wrap = FALSE))
}

