% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patchsizes.R
\name{patchsizes}
\alias{patchsizes}
\title{Get patch sizes.}
\usage{
patchsizes(x)
}
\arguments{
\item{x}{A binary matrix or a list of binary matrices.}
}
\value{
A vector of patch sizes or a list of vectors if the input was a list
  of binary matrices.
}
\description{
Get the distribution of patch sizes
}
\examples{
data(B)
patchsizes(B)
}

