% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchsizes.R
\name{patchsizes}
\alias{patchsizes}
\title{Get patch sizes.}
\usage{
patchsizes(x)
}
\arguments{
\item{x}{A binary matrix or a list of binary matrices.}
}
\value{
A vector of patch sizes or a list of vectors if the input was a list
  of binary matrices. If input contained no patches (all values to FALSE), 
  then NA is returned.
}
\description{
Get the distribution of patch sizes
}
\examples{
data(forestdat)
patchsizes(forestdat[['matrices']][[1]])

}

