% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_spectral_indic_plot.R
\name{plot.spectral_spews_test}
\alias{plot.spectral_spews_test}
\title{Plot method for Spectral Density Ratio (SDR)}
\usage{
plot.spectral_spews_test(obj, along = NULL, what = "value",
  display_null = TRUE)
}
\arguments{
\item{obj}{An \code{spectral_spews_test} object as produced by \link{indictest}}

\item{along}{An optional vector of values along which the trend is to be 
displayed. The length of the vector must be equal to the number of 
indicator values provided in the object \code{obj}.}

\item{what}{The value to plot on the y-axis. It defaults to the SDR value 
but alternative values can be \code{'z_score'} or \code{'pval'}.}

\item{display_null}{Whether to display the null-distribution ribbon}
}
\description{
This method reads the Spectral Density Ratio (SDR) values 
  produced by \code{\link{indictest}} and plots the trend with null values.
}
\examples{

data(forestdat)
sdr_test <- indictest(spectral_spews(forestdat[['matrices']]))
plot(sdr_test)

# Display the z_score along the vector of parameter values
plot(sdr_test, 
     along = forestdat[['parameters']][ ,'delta'], 
     what = 'z_score')

}

