% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raw_moran}
\alias{raw_moran}
\title{Compute the Moran's I at lag 1}
\usage{
raw_moran(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
The Moran's I numeric value as a numeric number.
}
\description{
This function computes the Moran'I value at lag 1.
}
\details{
See \link{indicator_moran} or \link{generic_spews} 
  for more information
}
\seealso{
\link{indicator_moran}, \link{generic_spews}
}
