% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raw_skewness}
\alias{raw_skewness}
\title{Skewness}
\usage{
raw_skewness(X)
}
\arguments{
\item{X}{A vector of values}
}
\description{
Compute the skewness of a given set of values
}
\seealso{
\code{\link[moments]{skewness}}
}
