% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_variogram_indic.R
\name{raw_structvar}
\alias{raw_structvar}
\title{Structural variance}
\usage{
raw_structvar(mat, subset_frac = 0.05, model = "sph")
}
\arguments{
\item{mat}{A matrix}

\item{subset_frac}{The number of matrix cells to consider (as a 
proportion of the total number in the matrix) to compute the variogram. 
A lower value here will speed up computations, at the expense of precision.}

\item{model}{The variogram model to use, either "sph" (for a spherical model)
or "exp" (for an exponential model)}
}
\description{
Compute the structural variance on a matrix.
}
\examples{

raw_structvar(serengeti[[5]])

}
\seealso{
raw_variogram_metrics, variogram_sews
}
