% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatialwarnings.R
\name{spatialwarnings}
\alias{spatialwarnings}
\title{Spatial Warnings Indicators}
\usage{
spatialwarnings(mat, indicators = list(variance =
  spatialwarnings::indicator_variance, skewness =
  spatialwarnings::indicator_skewness, moran = spatialwarnings::indicator_moran,
  psdfit = spatialwarnings::indicator_fitpsd, largestpatch =
  spatialwarnings::indicator_largestpatch, fracgeo =
  spatialwarnings::indicator_fracgeo, cumpsd =
  spatialwarnings::indicator_cumpsd), args = list(), verbose = FALSE, ...)
}
\arguments{
\item{mat}{A binary matrix or a list of square binary matrices}

\item{indicators}{A list of indicator_* functions}

\item{args}{A named list to pass arguments to individual indicator functions}

\item{verbose}{Print debugging messages ?}
}
\description{
This functions computes all the indicators of critical
  transitions in a convenient form.
}
\references{
Kefi, S., Guttal, V., Brock, W. A., Carpenter, S. R., Ellison,
A. M., Livina, V. N., ... & Dakos, V. (2014). Early warning signals of
ecological transitions: methods for spatial patterns. PloS one, 9(3), e92097.
}

